% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RcppMixtureModel-package}
\alias{RcppMixtureModel-package}
\title{RcppMixtureModel: Implementation of Mixture Model using MCMC}
\description{
RcppMixtureModel: Implementation of Mixture Model using MCMC
}

